Change functions using RX_DATA.indication PDU_Length to expect 32 bits

+++ openair2/LAYER2/NR_MAC_gNB/mac_proto.h
@@ -343,7 +343,7 @@ void nr_rx_sdu(const module_id_t gnb_mod_idP,
                const sub_frame_t subframeP,
                const rnti_t rntiP,
                uint8_t * sduP,
-               const uint16_t sdu_lenP,
+               const uint32_t sdu_lenP,
                const uint16_t timing_advance,
                const uint8_t ul_cqi,
                const uint16_t rssi);
